Another quick commit
I didn't properly explain before that stores is being properly
displayed, links are being shown, and rows of deals are now on display.
Duplicate errors in the database has also been handled by updating the
repository and now I am making a css file to handle the design of the
gamedetail page with more accuracy because I currently do not like how
the information is cluttered together.